Remove dolfinx_utils
