+++ python/.isort.cfg
@@ -1,6 +1,6 @@
 [settings]
-src_paths = ["demo", "dolfinx", "dolfinx_utils", "test"]
-known_first_party = basix,dolfinx_utils,dolfinx,ffcx,ufl
+src_paths = ["demo", "dolfinx", "test"]
+known_first_party = basix,dolfinx,ffcx,ufl
 known_third_party = gmsh,numpy,pytest
 known_mpi = mpi4py,petsc4py
 sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,MPI,LOCALFOLDER
