+++ python/setup.py
@@ -71,8 +71,7 @@ setup(name='fenics-dolfinx',
       long_description='',
       packages=["dolfinx",
                 "dolfinx.fem",
-                "dolfinx.wrappers",
-                "dolfinx_utils.test"],
+                "dolfinx.wrappers"],
       package_data={'dolfinx.wrappers': ['*.h']},
       ext_modules=[CMakeExtension('dolfinx.cpp')],
       cmdclass=dict(build_ext=CMakeBuild),
