+++ python/test/unit/common/test_mpi.py
@@ -16,7 +16,6 @@ import dolfinx.pkgconfig
 from dolfinx import wrappers
 from dolfinx.jit import mpi_jit_decorator
 from dolfinx.mesh import create_unit_square
-from dolfinx_utils.test.fixtures import tempdir  # noqa: F401
 
 import mpi4py
 from mpi4py import MPI
