+++ python/test/unit/fem/test_assemble_cppimport.py
@@ -20,7 +20,6 @@ from dolfinx.fem import (FunctionSpace, assemble_matrix, dirichletbc, form,
                          locate_dofs_geometrical)
 from dolfinx.mesh import create_unit_square
 from dolfinx.wrappers import get_include_path as pybind_inc
-from dolfinx_utils.test.fixtures import tempdir  # noqa: F401
 
 import petsc4py
 from mpi4py import MPI
