+++ python/test/unit/io/test_xdmf_mesh.py
@@ -15,11 +15,8 @@ from dolfinx.io import XDMFFile, ufl_mesh_from_gmsh
 from dolfinx.mesh import (CellType, GhostMode, create_mesh, create_submesh,
                           create_unit_cube, create_unit_interval,
                           create_unit_square, locate_entities)
-from dolfinx_utils.test.fixtures import tempdir
 
 from mpi4py import MPI
-
-assert (tempdir)
 
 # Supported XDMF file encoding
 if MPI.COMM_WORLD.size > 1:
