+++ python/test/unit/io/test_xdmf_meshdata.py
@@ -11,11 +11,8 @@ import pytest
 from dolfinx.io import XDMFFile
 from dolfinx.mesh import (CellType, create_unit_cube, create_unit_interval,
                           create_unit_square)
-from dolfinx_utils.test.fixtures import tempdir
 
 from mpi4py import MPI
-
-assert (tempdir)
 
 # Supported XDMF file encoding
 if MPI.COMM_WORLD.size > 1:
