+++ python/test/unit/mesh/test_mesh_partitioners.py
@@ -15,11 +15,8 @@ import ufl
 from dolfinx.io import XDMFFile
 from dolfinx.mesh import (CellType, GhostMode, compute_midpoints, create_box,
                           create_cell_partitioner, create_mesh)
-from dolfinx_utils.test.fixtures import tempdir
 
 from mpi4py import MPI
-
-assert (tempdir)
 
 partitioners = [dolfinx.graph.partitioner()]
 try:
